final 1.0

+++ src/app/Admin/Poste/getpostebyid/full-calendar-rh/full-calendar-rh.component.ts
@@ -8,6 +8,7 @@ import { PosteService } from 'src/app/service/poste.service';
 import dayGridPlugin from '@fullcalendar/daygrid';
 import timeGridPlugin from '@fullcalendar/timeGrid';
 import { UserService } from 'src/app/service/user.service';
+import Swal from 'sweetalert2';
 @Component({
   selector: 'app-full-calendar-rh',
   templateUrl: './full-calendar-rh.component.html',
@@ -49,7 +50,7 @@ export class FullCalendarRhComponent implements OnInit {
       this.dateEntretien = this.entretienDetails.entretien.dateEntretien;
       this.heureDebut = this.entretienDetails.entretien.heureDebut;
       this.heureFin = this.entretienDetails.entretien.heureFin;
-      this.userId = this.entretienDetails.entretien.userId; // Assuming entretien has userId
+      this.userId = this.entretienDetails.entretien.user.id; // Assuming entretien has userId
     }
     console.log('r', this.candidatureId);
     console.log('ra', this.entretienDetails);
@@ -159,16 +160,38 @@ export class FullCalendarRhComponent implements OnInit {
   }
 
   deleteEntretien(id: number): void {
-    this.entretienService.deleteEntretien(id).subscribe(
-      () => {
-        console.log('Entretien supprimé avec succès.');
-        // Ajoutez ici toute logique supplémentaire après la suppression réussie
-      },
-      (error) => {
-        console.error("Erreur lors de la suppression de l'entretien :", error);
-        // Gérer l'erreur ici
+    Swal.fire({
+      title: 'Êtes-vous sûr?',
+      text: "Vous ne pourrez pas revenir en arrière!",
+      icon: 'warning',
+      showCancelButton: true,
+      confirmButtonColor: '#3085d6',
+      cancelButtonColor: '#d33',
+      confirmButtonText: 'Oui, supprimez-le!'
+    }).then((result) => {
+      if (result.isConfirmed) {
+        this.entretienService.deleteEntretien(id).subscribe(
+          () => {
+            Swal.fire(
+              'Supprimé!',
+              "L'entretien a été supprimé.",
+              'success'
+            );
+            console.log('Entretien supprimé avec succès.');
+            // Ajoutez ici toute logique supplémentaire après la suppression réussie
+          },
+          (error) => {
+            Swal.fire(
+              'Erreur!',
+              "Erreur lors de la suppression de l'entretien: " + error.message,
+              'error'
+            );
+            console.error("Erreur lors de la suppression de l'entretien :", error);
+            // Gérer l'erreur ici
+          }
+        );
       }
-    );
+    });
   }
   editMode: boolean = false; // Variable pour suivre l'état de l'affichage du formulaire de modification
 
@@ -178,7 +201,6 @@ export class FullCalendarRhComponent implements OnInit {
     this.entretienService
       .updateEntretien(
         id,
-        this.candidatureId,
         this.dateEntretien,
         this.heureDebut,
         this.heureFin,
@@ -199,7 +221,6 @@ export class FullCalendarRhComponent implements OnInit {
     this.entretienService
       .updateEntretien(
         id,
-        this.candidatureId,
         this.dateEntretien,
         this.heureDebut,
         this.heureFin,
@@ -208,11 +229,24 @@ export class FullCalendarRhComponent implements OnInit {
       .subscribe(
         (response) => {
           console.log(response); // Afficher la réponse du serveur après la mise à jour de l'entretien
+          // Afficher un message de succès
+          Swal.fire({
+            icon: 'success',
+            title: 'Entretien mis à jour avec succès!',
+            showConfirmButton: false,
+            timer: 1500 // Durée pendant laquelle le message sera affiché (en millisecondes)
+          });
           // Réinitialiser le mode d'édition et recharger les données si nécessaire
           this.editMode = false;
         },
         (error) => {
           console.error("Erreur lors de la mise à jour de l'entretien:", error);
+          // Afficher un message d'erreur
+          Swal.fire({
+            icon: 'error',
+            title: 'Erreur',
+            text: 'Une erreur s\'est produite lors de la mise à jour de l\'entretien. Veuillez réessayer plus tard.',
+          });
           // Gérer l'erreur, afficher un message à l'utilisateur, etc.
         }
       );
